make uploaded shapefile and geojson editable with geoman, improve geoman handles
